requirements.txt
django
djangorestframework
python-dotenv
Pillow
gunicorn
whitenoise
pyyaml 
ua-parser 
dj_database_url
psycopg
django-storages
boto3
django_user_agents
pandas
django-cors-headers
django_filter
psycopg; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"